+++ README.md
@@ -30,6 +30,20 @@ make fix-formatting
 make test-all
 ```
 
+## Design decisions and structure
+With this repository we provide already pre-trained retina models that can be used for inference and intepretability out of the box, and dataloaders together with model architectures to train new models.
+For training new models, we rely on [pytorch lightning](https://lightning.ai/docs/pytorch/stable/) in combination with [hydra](https://hydra.cc/docs/intro/) to manage the configurations for training and dataloading.
+
+The openretina package is structured as follows:
+- modules: pytorch modules that define layers and losses
+- models: pytorch lightning models that define models that can be trained and evaluated
+- data_io: dataloaders to manage access of data to be used for training
+- model_anaysis: Methods to interpret above models
+    - stimulus_optimization: optimize inputs for neurons of above models according to interpretable objectives (e.g. most exciting inputs)
+    - ... (gradient analysis, data analysis)
+- utils: Utility functions that are used across above submodules
+
+
 ## Related papers
 
 The model in `openretina/hoefling_2024` was developed in the paper [A chromatic feature detector in the retina signals visual context changes](https://www.biorxiv.org/content/10.1101/2022.11.30.518492.abstract) and can be cited as:
