+++ openretina/data_io/cyclers.py
