+++ openretina/data_io/hoefling_2024/constants.py
