+++ openretina/data_io/hoefling_2024/nnfabrik_model_loading.py
