+++ openretina/data_io/stimuli.py
