+++ openretina/data_io/stimuli/RGC_Chirp_30Hz_18_16.pickle
