+++ openretina/data_io/stimuli/RGC_MovingBar_30Hz_18_16.pickle
