+++ openretina/legacy/early_stopping.py
