+++ openretina/model_analysis/stimulus_optimization/objective.py
