+++ openretina/model_analysis/stimulus_optimization/optimization_stopper.py
