+++ openretina/model_analysis/stimulus_optimization/regularizer.py
