+++ openretina/utils/plotting.py
@@ -14,9 +14,9 @@ from matplotlib import animation
 from matplotlib.colors import Normalize
 from matplotlib.patches import Rectangle
 
-from openretina.hoefling_2024.configs import MEAN_STD_DICT_74x64, pre_normalisation_values_18x16
-from openretina.hoefling_2024.constants import FRAME_RATE_MODEL
-from openretina.video_analysis import calculate_fft, decompose_kernel, weighted_main_frequency
+from openretina.legacy.configs import MEAN_STD_DICT_74x64, pre_normalisation_values_18x16
+from openretina.data_io.hoefling_2024 import FRAME_RATE_MODEL
+from openretina.utils.video_analysis import calculate_fft, decompose_kernel, weighted_main_frequency
 
 # Longer animations
 matplotlib.rcParams["animation.embed_limit"] = 2**128
