+++ scripts/train_core_readout.py
@@ -10,13 +10,13 @@ import torch
 from lightning.pytorch.callbacks import ModelCheckpoint
 from omegaconf import DictConfig, OmegaConf
 
-from openretina.cyclers import LongCycler
-from openretina.hoefling_2024.data_io import (
+from openretina.data_io.cyclers import LongCycler
+from openretina.data_io.hoefling_2024 import (
     natmov_dataloaders_v2,
 )
 from openretina.models.core_readout import CoreReadout
 from openretina.models.model_utils import OptimizerResetCallback
-from openretina.neuron_data_io import filter_responses, make_final_responses
+from openretina.data_io.neuron_data_io import filter_responses, make_final_responses
 from openretina.utils.h5_handling import load_h5_into_dict
